+++ teamBootstrap/src/main/java/ksmart39/springboot/controller/QualityControlController.java
@@ -91,49 +91,7 @@ public class QualityControlController {
 		return"quality/addFinalnspectionMeasurementValue";
 	}
 	
-	
-	
 
-	//[보람] 검사 리스트 검사번호클릭시 검사정보 경로
-			@GetMapping("qualityInspectionInfo")
-			public String qualityInspectionInfo() {
-				return "quality/qualityInspectionInfo";
-			}
-			
-	//[보람 ]검사 수정 완료
-	@PostMapping("/modifyQualityInspection")
-	public String  modifyQualityInspection() {
-		
-		return"redirect:/qualityInspectionList";
-	}
-	//[보람] 검사 수정 경로
-	@GetMapping("/modifyQualityInspection")
-	public String getModifyQualityInspection() {
-		return "quality/modifyQualityInspection";
-	}
-	
-	
-	//검사종류 리스트 메서드
-	@GetMapping("/qualityInspectionList")
-	public String getQualityInspectionList(Model model) {
-		List<QualityInspection> qualityInspectionList =qualityControlService.getQualityInspectionList();
-		
-		log.info("========================================");
-		log.info("qualityInspectionList:",qualityInspectionList);
-		log.info("========================================");
-		model.addAttribute("qualityInspectionList", qualityInspectionList);
-	
-		
-		return"quality/qualityInspectionList";
-	}
-	
-	//검사종류 등록 메서드
-	@GetMapping("/addQualityInspection")
-	public String addQualityInspection(Model model) {
-		
-		model.addAttribute("title", "품질검사:검사등록");
-		return"quality/addQualityInspection";
-	}	
 	
 
 	@GetMapping("/addStandardTable")
