+++ teamBootstrap/src/main/java/ksmart39/springboot/service/SystemService.java
@@ -1,5 +1,39 @@
 package ksmart39.springboot.service;
 
-public class SystemService {
+import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ksmart39.springboot.dao.SystemMapper;
+import ksmart39.springboot.domain.AccountingCategory;
+import ksmart39.springboot.domain.HumanResources;
+import ksmart39.springboot.domain.QualityInspection;
+
+@Service
+public class SystemService {
+	
+	@Autowired
+	private SystemMapper systemMapper;
+
+	
+	public List<AccountingCategory> getAccountingSubjectList(){		
+		return systemMapper.getAccountingSubjectList();
+		
+	}
+	
+	//======================================
+	//회원전체조회(levelName포함)
+	public List<HumanResources> getHumanResources(){
+		List<HumanResources> humanResources = systemMapper.getHumanResources();
+		return humanResources;
+	}
+	
+	//========================================
+	
+	//[보람 ]품질검사 리스트 조회메서드
+	public List<QualityInspection> getQualityInspectionList(){
+		List<QualityInspection> qualityInspectionList = systemMapper.getQualityInspectionList();
+	
+		return   qualityInspectionList;
+	}
 }
