--- teamBootstrap/src/main/resources/mapper/HumanResourcesMapper.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="ksmart39.springboot.dao.HumanResourcesMapper">
-	<resultMap type="HumanResources"  id="getHumanResourcesInfo">
-		<result  property="employeeCode" 				column="employee_code"/>
-		<result  property="levelNum" 				column="level_num"/>
-		<result  property="employeeId" 				column="employee_id"/>
-		<result  property="employeePw" 			column="employee_pw"/>
-		<result  property="employeeName" 			column="employee_name"/>
-		<result  property="employeeTel" 				column="employee_tel"/>
-		<result  property="employeeAddress" 				column="employee_address"/>
-		<result  property="employeeEmail" 				column="employee_email"/>
-		<result  property="employeeDepartment" 				column="employee_department"/>
-		<result  property="employeePosition" 		column="employee_position"/>
-		<result  property="employeeInfoRegDate" 		column="employee_info_reg_date"/>
-		<result  property="employeeInfoUpdateDate" 		column="employee_info_update_date"/>
-		<association property="humanResourcesLevel"  javaType="HumanResourcesLevel">
-			<id property="levelNum"  column="level_num"/>	 <!-- join 공통 -->
-			<result property="levelName" column="level_name"/>
-		</association>
-	</resultMap>
-	<select id="getHumanResources" resultMap="getHumanResourcesInfo"  resultType="HumanResources" fetchSize="1000">
-		SELECT
-			e.employee_code
-			,e.level_num
-			,e.employee_id
-			,e.employee_pw
-			,e.employee_name
-			,e.employee_tel
-			,e.employee_address
-			,e.employee_email
-			,e.employee_department
-			,e.employee_position
-			,e.employee_info_reg_date
-			,e.employee_info_update_date
-			,l.level_name
-		FROM
-			employee_info as e
-			INNER JOIN
-			erp_access_level as l
-			on
-			e.level_num = l.level_num;
-	</select>
-</mapper>
