--- teamBootstrap/src/main/resources/templates/accounting/accounting.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-layout:decorate="@{layout/default_accounting}">
-
-<th:block layout:fragment="pageTitle">
-	   <div class="border-head"> 
-	    <h3>결제 요청 등록</h3>
-	    </div>
-</th:block>
-
-
-<th:block layout:fragment="container">
-	회계, 전표 수정
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-</th:block>
-
-</html>
