--- teamBootstrap/src/main/resources/templates/community/community.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-layout:decorate="@{layout/default_community}">
-
-<th:block layout:fragment="pageTitle">
-	   <div class="border-head"> 
-	    <h3>시스템관리</h3>
-	    </div>
-</th:block>
-
-
-<th:block layout:fragment="container">
-	시스템관리 
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-	<br><br><br><br><br><br><br><br><br><br><br><br>
-</th:block>
-
-</html>
