+++ teamBootstrap/src/main/resources/templates/contract/addInvoiceApproval.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org"
 	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-	  layout:decorate="@{layout/default_accounting}">
+	  layout:decorate="@{layout/default_contract}">
 
 <th:block layout:fragment="pageTitle">
 	   <div class="border-head"><h3>결제 관리</h3></div>
