+++ teamBootstrap/src/main/resources/templates/contract/addPaymentInvoice.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org"
 	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-	  layout:decorate="@{layout/default_accounting}">
+	  layout:decorate="@{layout/default_contract}">
 
 <th:block layout:fragment="pageTitle">
 	   <div class="border-head">
