--- teamBootstrap/src/main/resources/templates/contract/buyerList.html
@@ -1,124 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-layout:decorate="@{layout/default_contract}">
-
-<th:block layout:fragment="pageTitle">
-	   <div class="border-head"> 
-	    <h3 th:text="${title}"></h3>
-	    </div>
-</th:block>
-
-<th:block layout:fragment="customJs">
-	<script>
-	//#checkAll 버튼 클릭 시 전체 checked
-		$('#checkAll').click(function(){
-			if($('#checkAll').prop('checked')){
-				$('.checked').prop('checked',true);
-					}else{
-						$('.checked').prop('checked',false);
-					}
-				});									
-			});
-
-</script>
-
-</th:block>
-
-<th:block layout:fragment="container">
-
-	
-	
-	<div class="row mt">
-     <div class="col-md-12">
-		<h4><i class="fa fa-angle-right"></i>수주거래처조회</h4>
-          <hr>  
-			<div class="form-group" style="padding-left: 50px">
-				<div class="col-md-3">
-				 <select class="form-control">
-				   <option>거래처명</option>
-				   <option>전화번호</option>
-				   <option>대표자명</option>
-				   <option>담당자명</option>
-				   <option>사업자번호</option>
-				   <option>주소</option>
-				   <option>등록일자</option>
-
-				  </select>
-				</div>
-				<div class="col-md-6">
-				  <input class="form-control" type="text" placeholder="검색어를 입력하세요">
-				</div>
-				<div class="col-md-2">
-				  <!-- 버튼 누르면 검색할 수 있는 창으로 연결 -->            
-				  <button type="button" class="btn btn-primary">검색</button>
-				</div>
-	       </div>
-	       
-	      <br>
-	      <br>
-	      <br>
-	      <br>
-	      <br>
-	      
-                
-	        <form class="form-horizontal style-form">
-	        	<h4><i class="fa fa-angle-right"></i>수주거래처목록</h4>
-	        
-              <table class="table table-striped table-advance table-hover" id="list">
-                <thead>
-                  <tr>
-                  	<th><input type="checkbox" id="checkAll"></th>
-                    <th><i class="fa fa-bullhorn"></i>번호</th>
-                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>거래처코드</th>
-                    <th><i class="fa fa-bookmark"></i>회사명</th>
-                    <th><i class=" fa fa-bookmark"></i>전화번호</th>
-                    <th><i class="fa fa-bookmark"></i>대표자</th>
-                    <th><i class=" fa fa-bookmark"></i>담당자</th>
-                    <th><i class=" fa fa-bookmark"></i>사업자번호</th>
-                    <th><i class=" fa fa-bookmark"></i>주소</th>
-                    <th><i class=" fa fa-bookmark"></i>등록일자</th>
-                    <th><i class=" fa fa-pencil"></i>수정</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                  	<td><input type="checkbox" id="checkAll"></td>
-                    <td class="hidden-phone">1</td>
-                    <td class="hidden-phone">buyer01</td>
-                    <td>이한빛</td>
-                    <td>063-293-9192</td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>조인성</td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>이광수</td>
-                    <td>109-203123-23124</td>
-                    <td>전라북도 전주시 만성동로 129</td>
-                    <td>2021.06.21 12:25</td>
-                  	<td>  <button class="btn btn-primary btn-xs" type="button" onclick="location.href='/modifyBuyer'"><i class="fa fa-pencil"></i>수정</button></td>
-
-                    </tr>
-                  
-               
-
-                 
-                 
-                </tbody>
-              </table>
-            
-		        <div class="form-group text-right">
-		          <div class="col-lg-offset-2 col-lg-10">
-		          <a th:href="@{/addBuyer}" class="btn btn-success btn-xs"><i class="fa fa-plus"></i>거래처등록</a>	          
-		           <a id="deleteBtn" class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i>삭제</a>
-		          </div>
-		        </div>
-       		</form>
-       		
-          </div>
-          <!-- /col-md-12 -->
-        </div>
-        <!-- /row -->
-        
-</th:block>
-
-</html>
