--- teamBootstrap/src/main/resources/templates/contract/supplierList.html
@@ -1,198 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-layout:decorate="@{layout/default_contract}">
-
-<th:block layout:fragment="pageTitle">
-	   <div class="border-head"> 
-	   <h3></h3>
-	    </div>
-	    
-<th:block layout:fragment="customJs">
-<script>
-$(function(){
-	//#checkAll 버튼 클릭 시 전체 checked
-		$('#checkAll').click(function(){
-			if($('#checkAll').prop('checked')){
-				$('.check').prop('checked',true);
-					}else{
-						$('.check').prop('checked',false);
-					}
-				});
-
-		//삭제하기
-		var tableAttr = $('#supplierList').children();
-		var tdList = tableAttr.find('.check');
-		
-		$('#deleteBtn').click(function(){
-			tdList = tableAttr.find('.check');
-			var parent = tdList.parent();
-			parent.each(function(){
-				console.log($(this).parent());	
-				var removeRow = $(this).parent();			
-				if($(this).find('.check').is(':checked')){		
-					removeRow.remove();
-				}
-			});
-
-		});
-			});
-
-
-</script>
-</th:block>
-
-
-<th:block layout:fragment="container">
-     <div class="col-md-12">
-       <div class="content">
-       
-        <h4><i class="fa fa-angle-right"></i>발주거래처조회</h4>
-         	 <hr>
-		<form class="form-inline" role="form">
-        
-         <table class="table table-striped table-advance table-hover"  >
-         	
-         	<tbody>
-         		<tr>
-         			<td style="padding-right: 20px;">
-	         			 <div class="form-group  col-md-5"><!-- 발주처별 -->
-		               		 <select class="form-control">
-				   				<option>발주거래처명</option>
-				   				<option>발주거래처담당자명</option>
-				   				<option>발주거래처대표자명</option>
-				   				<option>발주거래처사업자번호</option>
-				   			 </select>
-		                </div>
-         			</td>
-         			<td >
-         				<div class="form-group col-md-5 ">
-		                	<input type=text class="form-control"  placeholder="소분류검사명입력하세요" style="width: 300px;">
-	               		</div>
-         			</td>
-         			<td style="padding-right:100px;">
-         				  <button type="button" class="btn btn-theme">검색</button>
-         			</td>        			
-         			
-         			
-         			
-         		</tr>
-         	</tbody>
-         </table>  
-        </form>  
-	      
-              <h4><i class="fa fa-angle-right"></i>발주거래처 목록</h4>    
-              <table class="table table-striped table-advance table-hover" id="supplierList">
-                <thead>
-                  <tr>
-                    <th><input type="checkbox" class="check" id="checkAll"></th>
-                    <th><i class="fa fa-bullhorn"></i>순번</th>
-                    <th><i class="fa fa-bullhorn"></i>거래처코드</th>
-                    <th><i class="fa fa-bookmark"></i>거래처명</th>
-                    <th><i class=" fa fa-bookmark"></i>거래처담당자</th>
-                    <th><i class=" fa fa-bookmark"></i>거래처연락처</th>
-                  	<th><i class=" fa fa-bookmark"></i>거래처대표자명</th>
-                    <th><i class=" fa fa-bookmark"></i>거래처사업자번호</th>
-                    <th><i class=" fa fa-bookmark"></i> 거래처주소 </th>
-                    <th><i class=" fa fa-bookmark"></i>발주거래처등록일</th>
-                    <th><i class=" fa fa-bookmark"></i> 수정</th>
-                                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                   <td class="hidden-phone">
-                      <input type="checkbox" class="check">
-                    </td>
-                    <td class="hidden-phone">1</td>
-                    <td class="hidden-phone"  id="supplierCode">1028</td>
-                    <td class="hidden-phone" id="supplierName">나는단추</td>
-                    <td id="supplierEmployee">홍길동</td>
-                    <td id="supplierNumber">070-2515-1555</td>
-                    <td id="supplierOner">김철수</td>
-                    <td id="supplier1">111-1011-111</td>
-                    <td id="suppelirAddr">
-                     	전주시
-                    </td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>2021-01-17 11:30</td>
-                  	<td>  <button class="btn btn-primary btn-xs" type="button" id="modifyBtn" onclick="location.herf='/modifySupplier'"><i class="fa fa-pencil"></i>수정</button></td>
-                  </tr>
-                  
-                  <tr>
-                   <td class="hidden-phone">
-                      <input type="checkbox" class="check">
-                    </td>
-                    <td class="hidden-phone">2</td>
-                    <td class="hidden-phone" id="supplierCode">1029</td>
-                   <td class="hidden-phone">견고한지퍼나라</td>
-                    <td>공예윤</td>
-                    <td>070-2505-1545</td>
-                    <td>공예찬</td>
-                    <td>403-81-27365</td>
-                    <td>
-                     	익산시
-                    </td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>2021-01-18 11:30</td>
-                  	<td>  <button class="btn btn-primary btn-xs" type="button"onclick="location.herf='/modifySupplier'"><i class="fa fa-pencil"></i>수정</button></td>
-                  </tr>
-                  
-                  <tr>
-                   <td class="hidden-phone">
-                      <input type="checkbox" class="check">
-                    </td>
-                    <td class="hidden-phone">3</td>
-                    <td class="hidden-phone" id="supplierCode">1030</td>
-                    <td class="hidden-phone">부드러운실크컴퍼니</td>
-                    <td>서윤우</td>
-                    <td>070-2815-1555</td>
-                    <td>김아현</td>
-                    <td>403-29-87654</td>
-                    <td>
-                     	김제시
-                    </td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>2021-01-19 11:30</td>
-                  	<td>  <button class="btn btn-primary btn-xs" type="button" onclick="location.herf='/modifySupplier'"><i class="fa fa-pencil"></i>수정</button></td>
-                  </tr>
-                  
-                  <tr>
-                   <td class="hidden-phone">
-                      <input type="checkbox" class="check">
-                    </td>
-                    <td class="hidden-phone">4</td>
-                    <td class="hidden-phone" id="supplierCode">1031</td>
-                   <td class="hidden-phone">정교한심지</td>
-                    <td>이로운</td>
-                    <td>070-658-9208</td>
-                    <td>윤태웅</td>
-                    <td>413-05-89045</td>
-                    <td>
-                     	군산시
-                    </td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>2021-01-1811:30</td>
-                  	<td>  <button class="btn btn-primary btn-xs" type="button" onclick="location.herf='/modifySupplier'"><i class="fa fa-pencil"></i>수정</button></td>
-                  </tr>
-           		 </tbody>
-              </table>
-            </div>
-            <!-- /content-panel -->
-          </div>
-          	 <div class="form-group">
-                    <div class="col-lg-offset-1 col-lg-11 text-right">
-                      <button class="btn btn-success btn-xs" type="button" onclick="location.href='/addSupplier'"><i class="fa fa-check"></i>거래처등록</button>
-                      <button class="btn btn-danger btn-xs" type="button" id="deleteBtn"><i class="fa fa-trash-o "></i>거래처삭제</button>
-                    </div>
-                    </div>
-          <!-- /col-md-12 -->
-
-          <!-- ========================================================== -->
-        
-
-			
-						 </script>
-          
-</th:block>
-
-</html>
