+++ teamBootstrap/src/main/resources/templates/fragments/header.html
@@ -2,7 +2,7 @@
 <html xmlns:th="http://www.thymeleaf.org">
     <header th:fragment="header" class="header black-bg">
       <!--logo start-->
-      <a th:href="@{/}" class="logo"><b>TEAMONE<span>ERP</span></b></a>
+      <a th:href="@{/}" class="logo"><b>TEAMONE<span>MES</span></b></a>
       <!--logo end-->
       <div class="nav notify-row" id="top_menu">
         <!--  notification start -->
@@ -26,17 +26,8 @@
         	<span class="menu">5.출하관리</span>
         </a>
 
-        <a th:href="@{/accounting}">
-        	<span class="menu">6.회계관리</span>
-        </a>
-        <a th:href="@{/humanResources}">
-        	<span class="menu">7.인사관리</span>
-        </a>
-        <a th:href="@{/community}">
-        	<span class="menu">커뮤니티</span>
-        </a>
         <a th:href="@{/system}">
-        	<span class="menu">8.시스템관리</span>
+        	<span class="menu">6.시스템관리</span>
         </a>
 		<a th:href="@{#}">
         	<span class="menu">🕐배포일시:21-06-30 16:01:00🕐</span>
