+++ teamBootstrap/src/main/resources/templates/fragments/sidebar_contract.html
@@ -13,8 +13,8 @@
               <span>수주관리 ![한빛]</span>
               </a>
             <ul class="sub">
-              <li><a th:href="@{/addBuyer}">거래처 등록 ![한빛]</a></li>
-              <li><a th:href="@{/buyerList}">거래처 조회/목록 ![한빛]</a></li>
+<!--               <li><a th:href="@{/addBuyer}">거래처 등록 ![한빛]</a></li>
+              <li><a th:href="@{/buyerList}">거래처 조회/목록 ![한빛]</a></li> -->
               <li><a th:href="@{/addBuyerOrder}">주문서 등록 ![한빛]</a></li>
               <li><a th:href="@{/addBuyerOrder_Minah}">주문서 등록 ![민아버전]</a></li>
               <li><a th:href="@{/buyerOrderList}">주문서 조회/목록 ![한빛]</a></li>
@@ -30,12 +30,27 @@
               <span>발주관리 ![보람]</span>
               </a>
             <ul class="sub">
-              <li><a th:href="@{/addSupplier}">거래처등록 ![보람]</a></li>
-              <li><a th:href="@{/supplierList}">거래처 조회/목록 ![보람]</a></li>
+<!--               <li><a th:href="@{/addSupplier}">거래처등록 ![보람]</a></li>
+              <li><a th:href="@{/supplierList}">거래처 조회/목록 ![보람]</a></li> -->
               <li><a th:href="@{/addSupplierRequest}">발주등록 ![보람]</a></li>
               <li><a th:href="@{/supplierRequestList}">발주 조회/목록 ![보람]</a></li>
               <li><a th:href="@{/addSupplierContract}">발주계약등록 ![보람]</a></li>
               <li><a th:href="@{/supplierContractList}">발주계약조회/목록 ![보람]</a></li>
+            </ul>
+          </li>
+ 
+          <li class="sub-menu">
+            <a href="javascript:;">
+              <i class="fa fa-book"></i>
+              <span>회계관리 ![보람]</span>
+              </a>
+            <ul class="sub">
+<!--               <li><a th:href="@{/addSupplier}">거래처등록 ![보람]</a></li>
+              <li><a th:href="@{/supplierList}">거래처 조회/목록 ![보람]</a></li> -->
+       	   	<li><a th:href="@{/addInvoiceApproval}">결제 요청 등록 ![민아]</a></li>
+       	   	<li><a th:href="@{/invoiceApprovalList}">결제 현황 조회 ![민아]</a></li>
+              <li><a th:href="@{/addPaymentInvoice}">전표 등록 ![다미]</a></li>
+              <li><a th:href="@{/paymentInvoiceList}">전표 현황 조회 #[다미]</a></li>
             </ul>
           </li>
