--- teamBootstrap/src/main/resources/templates/fragments/sidebar_humanResources.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-      <div th:fragment="sidebar_humanResources" id="sidebar" class="nav-collapse ">
-        <!-- sidebar menu start-->
-        <ul class="sidebar-menu" id="nav-accordion">
-          <h5 class="centered">인사관리</h5>
-
-
-          <li class="sub-menu">
-            <a href="javascript:;">
-              <i class="fa fa-cogs"></i>
-              <span>사원등록 ![한빛]</span>
-              </a>
-            <ul class="sub">
-              <li><a th:href="@{/addHumanResources}">사원등록 ![한빛]</a></li>
-            </ul>
-          </li>
-          <li class="sub-menu">
-            <a href="javascript:;">
-              <i class="fa fa-book"></i>
-              <span>사원조회 ![한빛]</span>
-              </a>
-            <ul class="sub">
-              <li><a th:href="@{/humanResourcesList}">조회 ![한빛]</a></li>
-            </ul>
-          </li>
- 
-    
-        </ul>
-        <!-- sidebar menu end-->
-      </div>
-</html>
