+++ teamBootstrap/src/main/resources/templates/fragments/sidebar_qualityControl.html
@@ -6,7 +6,7 @@
           <h5 class="centered">품질관리</h5>
 
 
-          <li class="sub-menu">
+<!--           <li class="sub-menu">
             <a href="javascript:;">
               <i class="fa fa-cogs"></i>
               <span>검사목록 ![보람]</span>
@@ -15,7 +15,7 @@
               <li><a th:href="@{/addQualityInspection}">검사 등록 ![보람]</a></li>
               <li><a th:href="@{/qualityInspectionList}">검사 목록/조회 ![보람]</a></li>
             </ul>
-          </li>
+          </li> -->
           <li class="sub-menu">
             <a href="javascript:;">
               <i class="fa fa-book"></i>
