+++ teamBootstrap/src/main/resources/templates/fragments/sidebar_rawMaterials.html
@@ -9,19 +9,6 @@
           
           <li class="sub-menu">
             <a href="javascript:;">
-              <i class="fa fa-th"></i>
-              <span>원부자재 정보 ![민아]</span>
-              </a>
-            <ul class="sub">
-              <li><a th:href="@{/addRawMaterials}">원부자재 등록 ![민아]</a></li>
-              <li><a th:href="@{/rawMaterialsList}">원부자재 조회 ![민아]</a></li>
-              <li><a th:href="@{/inventoryStatus}">원부자재 현재고 현황 ![민아]</a></li>
-              <li><a th:href="@{/warehousingList}">입/출고 종합 조회 ![민아]</a></li>
-              
-            </ul>
-          </li>
-          <li class="sub-menu">
-            <a href="javascript:;">
               <i class="fa fa-sign-in"></i>
               <span>자재입고 ![다미]</span>
               </a>
@@ -42,11 +29,22 @@
               <li><a th:href="@{/addExWarehousing}">출고등록  #[한빛]</a></li>
               <li><a th:href="@{/exWarehousingList}">출고현황  #[한빛]</a></li>
 
+
+        </ul>
+          <li class="sub-menu">
+            <a href="javascript:;">
+              <i class="fa fa-th"></i>
+              <span>자재현황 ![민아]</span>
+              </a>
+            <ul class="sub">
+              <li><a th:href="@{/inventoryStatus}">원부자재 현재고 현황 ![민아]</a></li>
+              <li><a th:href="@{/warehousingList}">입/출고 종합 조회 ![민아]</a></li>
+              
+            </ul>
+          </li>
               
 
             </ul>
-
-        </ul>
         <!-- sidebar menu end-->
       </div>
 </html>
