+++ teamBootstrap/src/main/resources/templates/system/clientList.html
@@ -11,14 +11,6 @@ layout:decorate="@{layout/default_system}">
 
 <th:block layout:fragment="customJs">
 	<script>
-		$(function(){
-			$('#deleteBtn').click(function(){
-				var tdList = $('#list').children().find('tr');
-				tdList.html('<span class="check"><input type="checkbox" class="checked"></span>');
-				console.log(tdList);
-			});
-		})
-		$(function(){
 	//#checkAll 버튼 클릭 시 전체 checked
 		$('#checkAll').click(function(){
 			if($('#checkAll').prop('checked')){
@@ -29,7 +21,8 @@ layout:decorate="@{layout/default_system}">
 				});									
 			});
 
-	</script>
+</script>
+
 </th:block>
 
 <th:block layout:fragment="container">
@@ -38,16 +31,18 @@ layout:decorate="@{layout/default_system}">
 	
 	<div class="row mt">
      <div class="col-md-12">
-          <h4><i class="fa fa-truck"></i>전체사원목록/조회</h4>
+		<h4><i class="fa fa-angle-right"></i>수주거래처조회</h4>
           <hr>  
 			<div class="form-group" style="padding-left: 50px">
 				<div class="col-md-3">
 				 <select class="form-control">
-				   <option>사원명</option>
-				   <option>담당부서</option>
-				   <option>담당직책</option>
-				   <option>이메일</option>
+				   <option>거래처명</option>
 				   <option>전화번호</option>
+				   <option>대표자명</option>
+				   <option>담당자명</option>
+				   <option>사업자번호</option>
+				   <option>주소</option>
+				   <option>등록일자</option>
 
 				  </select>
 				</div>
@@ -68,43 +63,52 @@ layout:decorate="@{layout/default_system}">
 	      
                 
 	        <form class="form-horizontal style-form">
+	        	<h4><i class="fa fa-angle-right"></i>수주거래처목록</h4>
+	        
               <table class="table table-striped table-advance table-hover" id="list">
                 <thead>
                   <tr>
                   	<th><input type="checkbox" id="checkAll"></th>
-                    <th><i class="fa fa-bullhorn"></i>no.</th>
-                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>사원코드</th>
-                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>사원명</th>
-                    <th><i class="fa fa-bookmark"></i>담당부서</th>
-                    <th><i class=" fa fa-bookmark"></i>담당직책</th>
+                    <th><i class="fa fa-bullhorn"></i>번호</th>
+                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>거래처코드</th>
+                    <th><i class="fa fa-bookmark"></i>회사명</th>
+                    <th><i class=" fa fa-bookmark"></i>전화번호</th>
+                    <th><i class="fa fa-bookmark"></i>대표자</th>
+                    <th><i class=" fa fa-bookmark"></i>담당자</th>
+                    <th><i class=" fa fa-bookmark"></i>사업자번호</th>
+                    <th><i class=" fa fa-bookmark"></i>주소</th>
                     <th><i class=" fa fa-bookmark"></i>등록일자</th>
-                    <th><i class=" fa fa-bookmark"></i>상세</th>
                     <th><i class=" fa fa-pencil"></i>수정</th>
                   </tr>
                 </thead>
                 <tbody>
                   <tr>
-                  	<td><input type="checkbox" class="checked"></td>
+                  	<td><input type="checkbox" id="checkAll"></td>
                     <td class="hidden-phone">1</td>
-                    <td>workder01</td>
+                    <td class="hidden-phone">buyer01</td>
                     <td>이한빛</td>
-                    <td>본사</td>
-                    <td>임원</td>
-                    <td>2021.06.22 14:20</td>
-                    <td class="hidden-phone"><a th:href="@{#}" data-popup-open="humanResourcesInfo" >정보확인</a></td>
-                    <td>  <button class="btn btn-primary btn-xs" type="button" onclick="location.href='/modifyHumanResources'"><i class="fa fa-pencil"></i>수정</button></td>
-                  </tr>
+                    <td>063-293-9192</td>
+                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
+                    <td>조인성</td>
+                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
+                    <td>이광수</td>
+                    <td>109-203123-23124</td>
+                    <td>전라북도 전주시 만성동로 129</td>
+                    <td>2021.06.21 12:25</td>
+                  	<td>  <button class="btn btn-primary btn-xs" type="button" onclick="location.href='/modifyClient'"><i class="fa fa-pencil"></i>수정</button></td>
+
+                    </tr>
                   
                
 
-                 s
+                 
                  
                 </tbody>
               </table>
             
 		        <div class="form-group text-right">
 		          <div class="col-lg-offset-2 col-lg-10">
-		           <a th:href="@{/addHumanResources}" class="btn btn-success btn-xs"><i class="fa fa-check"></i>사원등록</a>
+		          <a th:href="@{/addClient}" class="btn btn-success btn-xs"><i class="fa fa-plus"></i>거래처등록</a>	          
 		           <a id="deleteBtn" class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i>삭제</a>
 		          </div>
 		        </div>
