+++ teamBootstrap/src/main/resources/templates/system/humanResourcesList.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org"
 xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-layout:decorate="@{layout/default_humanResources}">
+layout:decorate="@{layout/default_system}">
 
 <th:block layout:fragment="pageTitle">
 	   <div class="border-head"> 
@@ -38,7 +38,7 @@ layout:decorate="@{layout/default_humanResources}">
 	
 	<div class="row mt">
      <div class="col-md-12">
-          <h4><i class="fa fa-truck"></i>사원목록/조회</h4>
+          <h4><i class="fa fa-truck"></i>사원조회</h4>
           <hr>  
 			<div class="form-group" style="padding-left: 50px">
 				<div class="col-md-3">
@@ -66,6 +66,7 @@ layout:decorate="@{layout/default_humanResources}">
 	      <br>
 	      <br>
 	      
+          <h4><i class="fa fa-angle-right"></i>사원 목록</h4>
                 
 	        <form class="form-horizontal style-form">
               <table class="table table-striped table-advance table-hover" id="list">
@@ -84,31 +85,25 @@ layout:decorate="@{layout/default_humanResources}">
                   </tr>
                 </thead>
                 <tbody>
-                  <tr>
-                  	<td><input type="checkbox" class="checked"></td>
-                    <td class="hidden-phone">1</td>
-                    <td class="hidden-phone">worker01</td>      
-                    <td class="hidden-phone">이한빛</td>
-                    <td>본사</td>
-                    <td>임원</td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>onelight@naver.com</td>
-                    <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
-                    <td>010-7170-7123</td>
-                    <td>2021.06.22 14:20</td>
+                  <tr th:each=" h,n : ${humanResources}">
+                  	<td><input type="checkbox"  class="checked"></td>
+                    <td th:text = ${n.index+1}></td>
+                    <td  th:text="${h.employeeCode}"></td>      
+                    <td th:text="${h.employeeName}"></td>
+                    <td th:text="${h.employeeDepartment}"></td>
+                    <td th:text="${h.employeePosition}"></td>
+                    <td th:text="${h.employeeEmail}"></td>
+                    <td th:text="${h.employeeTel}"></td>
+                    <td th:text="${h.employeeInfoRegDate}"></td>
                     <td>  <button class="btn btn-primary btn-xs" type="button" onclick="location.href='/modifyHumanResources'"><i class="fa fa-pencil"></i>수정</button></td>
                   </tr>
                   
-               
-
-                 
-                 
                 </tbody>
               </table>
             
 		        <div class="form-group text-right">
 		          <div class="col-lg-offset-2 col-lg-10">
-		          <a th:href="@{/addHumanResources}" class="btn btn-success btn-xs"><i class="fa fa-plus"></i>사원추가</a>	          
+		          <a th:href="@{/addHumanResources}" class="btn btn-success btn-xs"><i class="fa fa-plus"></i>사원등록</a>	          
 		           <a id="deleteBtn" class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i>삭제</a>
 		          </div>
 		        </div>
