+++ teamBootstrap/src/main/resources/templates/system/qualityInspectionList.html
@@ -1,0 +1,167 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
+layout:decorate="@{layout/default_system}">
+
+<th:block layout:fragment="pageTitle">
+	   <div class="border-head"> 
+	    <h3 ></h3>
+	    </div>
+<th:block layout:fragment="customJs">
+	<script>
+	$(function(){
+		//#checkAll 버튼 클릭 시 전체 checked
+			$('#checkAll').click(function(){
+				if($('#checkAll').prop('checked')){
+					$('.check').prop('checked',true);
+						}else{
+							$('.check').prop('checked',false);
+						}
+					});
+
+			//삭제하기
+			var tableAttr = $('#inspectionList').children();
+			var tdList = tableAttr.find('.check');
+			
+			$('#delBtn').click(function(){
+				tdList = tableAttr.find('.check');
+				var parent = tdList.parent();
+				parent.each(function(){
+					console.log($(this).parent());	
+					var removeRow = $(this).parent();			
+					if($(this).find('.check').is(':checked')){		
+						removeRow.remove();
+					}
+				});
+
+			});
+			
+			
+			
+			
+				});
+
+
+	</script>
+	</th:block>
+   	    
+</th:block>
+
+
+
+<th:block layout:fragment="container">
+	<div class="row mt">
+     <div class="col-md-12">
+       <div class="content">
+       
+          <h4><i class="fa fa-angle-right"></i>품질검사조회</h4>
+          <hr>
+          <form class="form-inline" role="form">
+        
+         <table class="table table-striped table-advance table-hover">
+         	<thead>
+         		<tr>
+         			<th>검사대분류명</th>
+         			<th>검사중분류명</th>
+         			<th>검사소분류명</th>
+         			<th><!-- 공백 --></th>
+         			
+         		</tr>
+         	</thead>
+         	<tbody>
+         		<tr>
+         			<td style="padding-right: 20px;">
+         				<div class="form-group">
+		                	<select class="form-control" style="width: 125px;" name="highClassificationName" id="highClassificationName" onchange="getSelectBox(this)">
+		                		<option value="">대분류 검사명</option>
+		                		<th:block th:each="qualityInspection : ${qualityInspectionList}">
+		                		<option th:value="${qualityInspection.highClassificationCode}" th:text="${qualityInspection.highClassificationName}"></option>
+		                		</th:block>
+		                		
+		                	</select>
+	               		</div>
+         			</td>
+         			<td style="padding-right: 20px;">
+         				<div class="form-group">
+		                	<select class="form-control" style="width: 125px;" name="mediumClassificationName" id="mediumClassificationName" onchange="getSelectBox(this)" >
+		                		<option value="">중분류검사명</option>
+		                		
+		                	
+		                	</select>
+	               		</div>
+         			</td>
+         			<td style="padding-right: 20px;">
+         				<div class="form-group">
+		                	<input type=text class="form-control"  placeholder="소분류검사명입력하세요">
+	               		</div>
+         			</td>
+	         	
+         			<td>
+         				  <button type="submit" class="btn btn-theme">검색</button>
+         			</td>        			
+         		</tr>
+         	</tbody>
+         </table>  
+        </form>  
+         
+          
+	      
+              <h4><i class="fa fa-angle-right"></i>품질검사목록</h4> 
+              <hr>   
+              <table id="inspectionList" class="table table-striped table-advance table-hover">
+                <thead>
+                  <tr>
+                   <th>
+                    	<input type="checkbox" class="check" id="checkAll">
+                    </th>
+                    <th><i class="fa fa-bullhorn"></i>순번</th>
+                    <th><i class="fa fa-bullhorn"></i>검사코드</th>
+                    <th><i class="fa fa-bookmark"></i>검사대분류명</th>
+                    <th><i class=" fa fa-bookmark"></i>검사중분류명</th>
+                    <th><i class=" fa fa-bookmark"></i>검사소분류명</th>
+                    <th><i class=" fa fa-bookmark"></i>검사 소분류상세명</th>
+                    <th><i class=" fa fa-bookmark"></i>등록날짜</th>
+                    <th><i class=" fa fa-bookmark"></i>기준표</th>
+                    
+                    <th><i class=" fa fa-bookmark"></i>수정버튼</th>
+                    <th></th>
+                  </tr>
+                </thead>
+                <tbody>
+                  <tr th:if="${#lists.size(qualityInspectionList) > 0}" th:each="l,i: ${qualityInspectionList}" >
+                   <td class="hidden-phone">
+                      <input type="checkbox" class="check" name="check">
+                    </td>
+                    <td th:text="${i.index+1}"><a id="infoBtn"></a></td>
+                    <td th:text="${l.qualityInspectionCode}"></td>
+                    <td th:text =${l.highClassificationName} id="highClassificatonName"></td>
+                    <td th:text =${l.mediumClassificationName} id="mediumClassification"></td>
+                    <td th:text="${l.lowClassificationName}" id="lowClassification"></td>
+                    <td th:text="${l.subClassificationName}" id="subClassification">
+                     
+                    </td>
+                     <td th:text="${l.qualityInspectionRegDate}" id="regDate">
+                     	
+                    </td>
+                   <td>  <button class="btn btn-warning btn-xs" type="button" onclick="location.href='/addStandardTable'"><i class="fa fa-pencil"></i>기준표등록</button></td>
+                   <td>  <button class="btn btn-primary btn-xs" type="button" ><i class="fa fa-pencil"></i>수정</button></td>
+                  </tr>
+                  
+                  
+                 
+           		 </tbody>
+              </table>
+            </div>
+             
+                    <div class=" col-xs-12 text-right">
+                   
+                      <button class="btn btn-success btn-xs" type="button"  onclick="location.href='/addQualityInspection'"><i class="fa fa-check"></i>검사등록</button>
+                      <button class="btn btn-info btn-xs" type="button" onclick="location.href='/qualityInspectionRequest'" ><i class="fa fa-pencil"></i>검사요청</button>
+                      <button class="btn btn-danger btn-xs" type="button" id="delBtn"><i class="fa fa-trash-o "></i>검사삭제</button>
+                    </div>
+                 
+            </div>
+            </div>
+</th:block>
+
+</html>
